ENH: Delete User API and added superAdmin roles

+++ pages/api/admin/deleteUser.js
@@ -11,13 +11,28 @@ const handler = async (req, res) => {
         }
 
         try {
-            const decoded = jwt.verify(token, process.env.JWT_SECRET); // Decode and verify token
+            const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+            // Only allow admins to delete users
             if (!decoded.isAdmin) {
                 return res.status(403).json({ error: "Access Denied: Admins only" });
             }
+
             const { id } = req.query;
+
+            // Check if the user to be deleted is a super admin
+            const userToDelete = await User.findById(id);
+            if (!userToDelete) {
+                return res.status(404).json({ error: "User not found" });
+            }
+
+            if (userToDelete.isSuperAdmin) {
+                return res.status(403).json({ error: "You cannot delete a super admin" });
+            }
+
             await User.findByIdAndDelete(id);
             res.status(200).json({ success: "User deleted successfully" });
+
         } catch (err) {
             res.status(400).json({ error: "Invalid token" });
         }
@@ -25,6 +40,5 @@ const handler = async (req, res) => {
         res.status(400).json({ error: "This method is not allowed" });
     }
 };
-
 
 export default connectDb(handler);
